tests-shared/requirements.txt
## Test Dependencies for MommyShops API

# Testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# HTTP testing
httpx==0.25.2
requests==2.31.0
requests-mock==1.11.0

# Test client
fastapi[all]==0.109.0
httpx==0.25.2

# Mocking
faker==22.0.0

# Image testing
Pillow==10.1.0

# Coverage reporting
coverage[toml]==7.3.4

